Add TicketCreateForm to application
A new Form named TicketCreateForm has been added to enable ticket creation. The form is injected as a singleton service and it's associated with a ViewModel. A new interface IModalContentControl was also added, used to setup modal window properties. The BacklogViewModel has been updated to use this form within a modal window.